refactor: extract SessionLibraryDateRange.swift from SessionLibraryQuery.swift
Splits the `SessionLibraryDateRange` struct (start/end date pair used
by the custom date filter) out of SessionLibraryQuery.swift.

Byte-identical move of lines 41-54. Both files import Foundation.

SessionLibraryQuery.swift: 61 → 46 lines. Tests: 667.
